refactor(compose): modular shell + Load — the Save round-trip completes
Structure (before lens packs grow it): main.rs is now just the entry
point; app.rs owns state + persistence; ui/* is one module per panel
(top_bar, status_bar, palette, inspector, charts, canvas, hal_window);
circuit.rs stays the engine, export.rs the only JSON path — now in
BOTH directions.

Load: from_world_model() inverts to_world_model() — Load button +
drag-a-.json-onto-the-window both restore a saved model onto the
canvas. Compose-only knobs ride in the model's extensible fields so
the trip is lossless: buffer release_rate in cognitive_params,
gradient conductance as a flow Parameter. Loaded named substances
become pickable. Non-compose-shaped models error with a reason
rather than silently dropping structure.

Round-trip test: every knob survives (kinds, names, rates, stocks,
release, substances, gradient mode + k) and the loaded circuit is
behaviorally identical over 30 ticks, conserving. 24 tests, clippy
clean.
