+++ bert-compose/Cargo.toml
@@ -10,3 +10,4 @@ eframe = "0.31"
 egui = "0.31"
 egui_plot = "=0.31.0"
 ureq = { version = "2.10", features = ["json"] }
+rfd = "0.15"
